backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
channels>=4.0
daphne>=4.0
channels-redis>=4.1
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
django-celery-beat>=2.5
